+++ src/app/dashboard/widgets/commands-widget/commands-widget.component.ts
@@ -77,6 +77,6 @@ export class CommandsWidgetComponent implements OnInit {
   }
 
   private ddMM(date: Date) {
-    return `${date.getDate()}/${(date.getMonth() + 1)}`;
+    return `${date.getDate().toString().padStart(2, '0')}/${(date.getMonth() + 1).toString().padStart(2, '0')}`;   
   }
 }