+++ src/app/messages-widget/messages-widget.component.ts
@@ -75,6 +75,6 @@ export class MessagesWidgetComponent implements OnInit {
   }
 
   private ddMM(date: Date) {
-    return `${date.getDate()}/${(date.getMonth() + 1)}`;
+    return `${date.getDate().toString().padStart(2, '0')}/${(date.getMonth() + 1).toString().padStart(2, '0')}`;   
   }
 }